Add Scores
Procedure to add the scores achieved by a Student in different tests viz. IELTS, TOEFL, GRE, GATE, GMAT, CAT, and CGPA which will be required to check eligibility against a College's requirements. The Procedure also derives the Year (number of years of undergraduate degree) attribute from the Enrolment and Graduation Dates stored in the Student table.